+++ test/sandbox/app/components/basic_object_slot_v2_component.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+class BasicObjectSlotV2Component < ViewComponent::Base
+  renders_one :link, "Link"
+
+  class Link < ViewComponent::Base
+    attr_reader :method
+
+    def initialize(method:)
+      @method = method
+    end
+  end
+
+  def call
+    helpers.tag.a(data: { method: link.method })
+  end
+end